+++ stories/components/data-table/tests/examples/multi-level-header.tsx
@@ -23,7 +23,7 @@ const Demo = () => {
 
         </DataTable.Head.Group >
         <DataTable.Head.Column name='keyword' children='Keyword' />
-        <DataTable.Head.Group borders={'both'} title={'Bprders both - Organic Sessions rganic Sessions rganic Sessions'}>
+        <DataTable.Head.Group borders={'both'} title={(<Ellipsis>Bprders both - Organic Sessions rganic Sessions rganic Sessions</Ellipsis>)}>
 
           <DataTable.Head.Column name='kd2' gtcWidth={'100px'} >
             <Ellipsis>