+++ semcore/drag-and-drop/CHANGELOG.md
@@ -8,6 +8,12 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 - `keyboardFocusEnhance` to css `:focus-visible` property.
 
+## [3.49.0] - 2025-04-11
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/card` [5.47.1 ~> 5.48.0]).
+
 ## [3.48.1] - 2025-04-04
 
 ### Changed